services/shared/python/mobility-common/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mobility-common"
version = "0.1.0"
description = "Shared models, events, and utilities for Smart Mobility Platform"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.1",
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
    "pyjwt>=2.8",
    "bcrypt>=4.1",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mobility_common"]
